Using auth middleware instead of having the auth logic in the controllers
-> Adding secretkey-auth middleware that throws an error if not a valid
request
-> Moving error handlers outside of the s3 controller
-> Adding start script in package.json
