+++ package.json
@@ -4,6 +4,7 @@
 	"description": "This module acts as a middleware server handles uploading/getting images to/from Amazon S3 server with different sizes.",
 	"main": "server.js",
 	"scripts": {
+		"start": "node server.js",
 		"test": "echo \"Error: no test specified\" && exit 1"
 	},
 	"repository": {
